services/tracker/requirements.txt
FlightRadarAPI>=1.3.0
redis>=5.0.0
python-dotenv>=1.0.0